Change DMARC policy for barrucadu.co.uk to "reject"
"none" was useful when first setting up DMARC, as I could check the
reports to make sure nothing was going wrong.  But this record has
been in place since mid-2019, I can send email just fine, and I get
occasional DMARC reports; so it's time to change it to "reject".