Allow backups user to publish to SNS topic
Right now my backups script needs credentials for both the backups and
notifications IAM users, which is due to how it used to work: backups
had its own secrets, and would call a notifications script which had
separate secrets.  But now that the backups script is calling AWS
directly, this separation is inconvenient.