[dns] Use TCP healthchecks for geoDNS
The default healthchecks connect via HTTPS and check for a 200 response
on a given path.  I haven't got that configured and don't really want
to, which means everything is unhealthy and the geoDNS falls back to the
default pool - all the IPs.

Instead, switch to a TCP healthcheck that just checks it can connect to
the server via port 443.